requirements.txt
numpy
matplotlib
scikit-learn
pillow
tqdm
